+++ component/processor.go
@@ -17,7 +17,6 @@ package component // import "go.opentelemetry.io/collector/component"
 import (
 	"context"
 
-	"go.opentelemetry.io/collector/component/componenterror"
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/consumer"
 )
@@ -122,7 +121,7 @@ func (f CreateTracesProcessorFunc) CreateTracesProcessor(
 	cfg config.Processor,
 	nextConsumer consumer.Traces) (TracesProcessor, error) {
 	if f == nil {
-		return nil, componenterror.ErrDataTypeIsNotSupported
+		return nil, ErrDataTypeIsNotSupported
 	}
 	return f(ctx, set, cfg, nextConsumer)
 }
@@ -138,7 +137,7 @@ func (f CreateMetricsProcessorFunc) CreateMetricsProcessor(
 	nextConsumer consumer.Metrics,
 ) (MetricsProcessor, error) {
 	if f == nil {
-		return nil, componenterror.ErrDataTypeIsNotSupported
+		return nil, ErrDataTypeIsNotSupported
 	}
 	return f(ctx, set, cfg, nextConsumer)
 }
@@ -154,7 +153,7 @@ func (f CreateLogsProcessorFunc) CreateLogsProcessor(
 	nextConsumer consumer.Logs,
 ) (LogsProcessor, error) {
 	if f == nil {
-		return nil, componenterror.ErrDataTypeIsNotSupported
+		return nil, ErrDataTypeIsNotSupported
 	}
 	return f(ctx, set, cfg, nextConsumer)
 }